pyproject.toml
[project]
name = "weather-alert"
version = "0.1.0"
description = "O WeatherAlert é um sistema de monitoramento de temperatura de localidades."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-celery-beat>=2.8.1",
    "django-ninja>=1.4.3",
    "httpx[http2]>=0.28.1",
    "loguru>=0.7.3",
    "python-decouple>=3.8",
    "redis[hiredis]>=6.2.0",
    "stamina>=25.1.0",
    "uvicorn[standard]>=0.34.3",
]

[dependency-groups]
dev = [
    "blue>=0.9.1",
    "isort>=6.0.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
    "respx>=0.22.0",
    "taskipy>=1.14.1",
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'
DJANGO_SETTINGS_MODULE = "weather_alert.settings"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["weatheralert", "locations", "temperature", "alerts", "integrations", "api"]
default_section = "THIRDPARTY"
skip_gitignore = true

[tool.taskipy.tasks]
format = "isort . && blue ."
run = "uvicorn weather_alert.asgi:application --host 0.0.0.0 --port 8000 --reload"
celery = "celery -A weather_alert worker --loglevel=info"
beat = "celery -A weather_alert beat --scheduler django_celery_beat.schedulers:DatabaseScheduler --loglevel=info"
test = 'pytest -x -vv'
